% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFile.R
\name{ExtractFile}
\alias{ExtractFile}
\title{ExtractFile}
\usage{
ExtractFile(files, n_file = 1, sequence)
}
\arguments{
\item{files}{a list of Risoe.BINfileData-class objects}

\item{n_file}{number of the item to be selected in the 'files' list}

\item{sequence}{selected sequence}
}
\value{
The Risoe.BINfileData-class object is rewritten to take into account only the data included in the sequence
}
\description{
selects the useful data from a Risoe.BINfileData-class object
}
