% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateau.plot.R
\name{plateau.plot}
\alias{plateau.plot}
\title{plateau.plot}
\usage{
plateau.plot(Figures, alpha = TRUE, Dose)
}
\arguments{
\item{Figures}{[list] (**required**) WinBUGS simulation result}

\item{alpha}{[logical] (**with default**)) True if alpha measurements included}

\item{Dose}{[numeric] or [list] (**required**): doses in seconds}
}
\value{
a plot of a preheat plateau test
}
\description{
plot a preheat plateau test
}
